Access to messages on console
