--- Collective/src/components/common/MessageList.js
@@ -1,34 +1,0 @@
-import React from 'react';
-import {
-    ListView, Text, Row, Image,
-    View, Subtitle, Caption
-} from '@shoutem/ui';
-import moment from 'moment';
-
-const Message = ({ msg }) => (
-    <Row>
-        <Image
-          styleName="small-avatar top"
-          source={{ uri: msg.author.avatar }}
-        />
-        <View styleName="vertical">
-            <View styleName="horizontal space-between">
-                <Subtitle>{msg.author.name}</Subtitle>
-                <Caption>{moment(msg.time).from(Date.now())}</Caption>
-            </View>
-            <Text styleName="multiline">{msg.text}</Text>
-        </View>
-    </Row>
-);
-
-
-const MessageList = ({ messages, onLayout }) => (
-    <ListView
-      data={messages}
-      autoHideHeader
-      renderRow={msg => <Message msg={msg} />}
-      onLayout={onLayout}
-    />
-);
-
-export default MessageList;
